Code

Here lie all of the docstrings in the GameBoy.jl codebase.

GameBoy.Button — Type

All of the buttons that can be pressed

GameBoy.Clock — Type

Main System Clock

GameBoy.Emulator — Type

The state of an emulator.

GameBoy.IORegisters — Type

Addresses of IO devices relative to 0xff00

Intentionally skipping sound since it is unimplemented in this emulator

GameBoy.Mmu — Type

Memory Mapping Unit Read and write bytes to the correct subsystem based on address.

GameBoy.Video — Type

Video subsystem

GameBoy.Component.reset! — Method

Power cycle the emulator

GameBoy.buttonstate! — Method

Directly set the state of a button.

GameBoy.doframe! — Method

Run one frame of emulation.

GameBoy.execute — Method

Press a button in the emulator.

GameBoy.ram! — Method

Restore cartridge RAM. Use ram to capture the state.

GameBoy.ram — Method

Grab a copy of the cartridge RAM. This captures the save state. Use ram! to restore this state.

GameBoy.read — Method

Read an arbitrary byte of memory

GameBoy.Processor.Cpu — Type

Main processor

GameBoy.Carts.Cartridge — Type

The place that stores game data. Traditionally stored on removable media to swap between different games.

GameBoy.DirectMemoryAccess.DMA — Type

Direct Memory Access metadata

GameBoy.RandomAccessMemory.Ram — Type

Random Access Memory.